app/backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["kohei3110 <kk.31108424@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
azure-ai-projects = "^1.0.0b4"
azure-identity = "^1.19.0"
python-multipart = "^0.0.20"
aiofiles = "^24.1.0"
jsonref = "^1.1.0"
azure-ai-ml = "^1.23.1"
autogen-agentchat = "^0.4.9.2"
autogen-ext = {extras = ["openai"], version = "^0.4.9.2"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
